move Bitstring to utils
